Add word `for`, add logic for realid papers etc.
The word `for` has been added in between paper title and year.
The date laid is taken into accout in the sort. So for papers with
the same title the one laid first will now appear first.
Relaid papers have a special treatment.
Update whatson calendar api url.
Change dates returned for a session to include period between
last session and this one.